check whether dir exists or not

+++ lib/rbgithook.rb
@@ -22,13 +22,17 @@ module Rbgithook
   def self.add(args)
     file_name = args[0]
     hook_command = args[1]
-    Dir.chdir(".rbgithook")
+    dir_name = ".rbgithook"
+    unless Dir.exist?(dir_name)
+      warning_message("Directory", file_name, hook_command)
+      exit 1
+    end
+    Dir.chdir(dir_name)
     if File.exist?(file_name)
       file = File.open(file_name.to_s, "a")
       file.write("\n#{hook_command}")
-      FileUtils.chmod(0o755, file_name)
     else
-      warn "File not found, please run `rbgithook set #{file_name} '#{hook_command}'`"
+      warning_message("File", file_name, hook_command)
     end
   end
 
@@ -46,5 +50,11 @@ module Rbgithook
       uninstall - Uninstall hook
       help   - Show this usage
     USAGE
+  end
+
+  private
+
+  def self.warning_message(target, file_name, hook_command)
+    warn "#{target} not found, please run `rbgithook set #{file_name} '#{hook_command}'`"
   end
 end
